% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrange_ggmatrix}
\alias{arrange_ggmatrix}
\title{Plot several ggmatrix objects}
\usage{
arrange_ggmatrix(m, lay)
}
\arguments{
\item{m}{List of ggmatrices}

\item{layout}{A table with numeric placement of plots in the grid}
}
\value{
gtable object to be plotted with grid.draw
}
\description{
Arrange ggmatrix objects by retrieving their gtables
and manually binding them with a specified layout.
Not much documentation on this stuff so mostly cobbled togheter
from SO:
}
\details{
\itemize{
\item https://stackoverflow.com/questions/17126128/ggplot2-does-not-appear-to-work-when-inside-a-function-r
\item https://stackoverflow.com/questions/49439982/ggplot-and-grob-arrangement
\item 
\itemize{
\item https://stackoverflow.com/questions/24331107/the-perils-of-aligning-plots-in-ggplot
}
\item 
\itemize{
\item https://stackoverflow.com/questions/27361364/set-the-height-of-the-graphs-y-axis-in-grid-arrange-but-not-of-the-entire-plot
}
\item https://stackoverflow.com/questions/35068129/arrange-ggplot-plots-grobs-with-same-widths-using-gtable-to-create-2x2-layout
}
}
