% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{avg_migr}
\alias{avg_migr}
\title{Find the average migration during a period}
\usage{
avg_migr(file, ancient, modern, g)
}
\arguments{
\item{file}{Path to the output file containing the migration information.}

\item{ancient}{More ancient (further from present) boundary of the period.}

\item{modern}{More modern (closer to present) boundary of the period.}

\item{g}{Generation time in years.}
}
\value{
A vector of average weighted mijs for each population. (Unsafe)
}
\description{
Take the sum of rate over given epochs and average by the number of generations.
}
