% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ne_at}
\alias{ne_at}
\alias{ne_at,smcsmc-method}
\alias{ne_at,msmc-method}
\title{Retrieve population size.}
\usage{
ne_at(object, ...)

\S4method{ne_at}{smcsmc}(object, time, g = 29)

\S4method{ne_at}{msmc}(object, time, g = 29, mu = 1.25e-08)
}
\arguments{
\item{time}{Time, in years, to retrieve}

\item{g}{Generation time in years}

\item{mu}{Mutation rate for MSMC scaling}
}
\value{
Vector of population sizes for each population, in the same order as the labels.
}
\description{
Fetches effective population size estimates at a particular time. Dispatches to
MSMC and SMC2 specific functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{smcsmc}: SMC2 implementation

\item \code{msmc}: MSMC implementation
}}

