% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{plot_migration_gif}
\alias{plot_migration_gif}
\title{Plot migration between two groups}
\usage{
plot_migration_gif(
  out,
  pop0 = "Pop1",
  pop1 = "Pop2",
  main = "",
  plot_diff = F,
  plotly = F,
  ylim = c(0, 0.00025),
  xlim = c(1000, 10000, 1e+05, 1e+06),
  labels = c("1", "10", "100", "1000"),
  t0 = 500,
  t1 = 1500000
)
}
\arguments{
\item{out}{Either i. A character path to an out file or ii. A character vector of paths.}

\item{pop0}{An optional label for Pop0. Defaults to 'Pop0'.}

\item{pop1}{An optional label for Pop1. Defaults to 'Pop1'}
}
\description{
Alternatively, plot migration between two sets of groups,
and average over both. This gives a geom_smooth with an
error measurement. This is useful for larger comparissons.
}
