% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{plot_ne}
\alias{plot_ne}
\title{Plot effective population size over time.}
\usage{
plot_ne(
  out,
  pop0 = "Pop1",
  pop1 = "Pop2",
  main = "",
  plot_diff = F,
  plotly = F,
  xlim = c(1000, 10000, 1e+05, 1e+06),
  labels = c("1", "10", "100", "1000"),
  ylim = c(6.5, 12),
  t0 = 1000,
  t1 = 1500000,
  return = F,
  previous = NULL
)
}
\arguments{
\item{out}{Either i. A character path to an out file or ii. A character vector of paths.}

\item{pop0}{An optional label for Pop0. Defaults to 'Pop0'.}

\item{pop1}{An optional label for Pop1. Defaults to 'Pop1'}
}
\description{
Plot effective population size over time.
}
